add tests for multi output lines. next up polygons

+++ util.hpp
@@ -6,8 +6,8 @@
 #include <stdlib.h>
 
 namespace util {
-	inline int double_to_fix(const double x, const double scale) {
-		return int((x) * scale + 0.4);
+	inline int double_to_fix(const double x, const int scale) {
+		return x * scale + 0.4;
 	}
 	inline double fix_to_double(const int x, const int scale) {
 		return (double)x / scale;
